requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
python-decouple==3.8
psycopg2-binary==2.9.9
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
sqlparse==0.5.1
Pillow==10.4.0
absl-py==2.1.0
astunparse==1.6.3
attrs==24.2.0
blinker==1.8.2
blosc2==2.5.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.1
gTTS==2.5.3
h5py==3.11.0
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.5
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
joblib==1.4.2
keras==3.5.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.15
ml-dtypes==0.4.0
msgpack==1.0.8
namex==0.0.8
ndindex==1.8
numexpr==2.10.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.4
py-cpuinfo==9.0.0
pycparser==2.22
pygame==2.6.0
Pygments==2.18.0
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
rich==13.8.0
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
sounddevice==0.5.0
tables==3.9.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.1

